Point SVMP components at their master branches
I'm starting to reorganize the way we do branching in the project.

Anything that doesn't strictly need a version-specific branch for
development, that is, anything that's not a fork of an upstream AOSP
repository, will be working off master. This will also get rid of the
naming ambiguity problems with branches and tags with the same name
conflicting.

The tagged version manifests will continue to exist, and can be used to
check out a specific release version of the ROM code using a command like:

  repo init -u https://github.com/SVMP/android -m svmp-1.2.xml

Otherwise, to get the latest development version:

  repo init -u https://github.com/SVMP/android

In both cases the "-b master" is implied.

Change-Id: I5d2df8f344009baec9e872e55561ffe6282c15cc